+++ src/components/logpane/logpane.jsx
@@ -19,6 +19,14 @@ import { FaRegCheckCircle } from "react-icons/fa";
 import { FaRegCircleXmark } from "react-icons/fa6";
 import Modal from "../../components/modal/Modal";
 
+import { getDiscoveryUrl, dhcpScanURL } from "../../utils/backend_rest_urls.js";
+import CredentialForm from "../../pages/ZTPnDHCP/CredentialsForm";
+
+import {
+  defaultColDef,
+  dhcpColumn,
+} from "../../components/tabbedpane/datatablesourse";
+
 export const LogViewer = () => {
   const logPannelDivRef = useRef(null);
 
@@ -206,6 +214,14 @@ export const LogViewer = () => {
 
   const [height, setHeight] = useState(400);
 
+  const [showDhcpTable, setShowDhcpTable] = useState(false);
+  const [heightDhcpTable, setHeightDhcpTable] = useState(250);
+  const [sshData, setSshData] = useState(false);
+  const [dhcpTable, setDhcpTable] = useState([]);
+  const [selectedRows, setSelectedRows] = useState([]);
+  const dhcpTableRef = useRef(null);
+  const gridRefDhcpTable = useRef();
+
   useEffect(() => {
     if (updateLog) {
       getLogs();
@@ -215,6 +231,11 @@ export const LogViewer = () => {
   useEffect(() => {
     getLogs();
   }, []);
+
+  useEffect(() => {
+    console.log(window.location.href.includes("/home"));
+    setShowDhcpTable(window.location.href.includes("/home"));
+  }, [window.location.href]);
 
   const getLogs = () => {
     setLogEntries([]);
@@ -316,123 +337,230 @@ export const LogViewer = () => {
       });
   };
 
+  const handleResizeDhcpTable = () => {
+    if (dhcpTableRef.current.offsetHeight > 250) {
+      setHeightDhcpTable(dhcpTableRef.current.offsetHeight);
+    }
+  };
+
+  const onSelectionChangedDhcp = () => {
+    const selectedNodes = gridRefDhcpTable.current.api.getSelectedNodes();
+    const selectedData = selectedNodes.map((node) => node.data.device_ip);
+    setSelectedRows(selectedData);
+  };
+
+  const discoverDhcp = async () => {
+    try {
+      const response = await instance.put(getDiscoveryUrl(), {
+        address: selectedRows,
+        discover_from_config: true,
+      });
+    } catch (error) {
+      console.error(error);
+    } finally {
+    }
+  };
+
+  const scanDhcp = () => {
+    instance
+      .put(dhcpScanURL(), {
+        mgt_ip: sshData.device_ip,
+      })
+      .then((res) => {})
+      .catch((err) => {
+        console.error(err);
+      })
+      .finally(() => {
+        getLogs();
+      });
+  };
+
+  const gridStyleDhcpTable = useMemo(
+    () => ({ height: heightDhcpTable + "px", width: "100%" }),
+    [heightDhcpTable]
+  );
+
   return (
-    <div
-      className="logPanel resizable"
-      id="logPanel"
-      ref={logPannelDivRef}
-      onMouseMove={handleResize}
-    >
-      <div
-        className="mb-15"
-        style={{
-          display: "flex",
-          justifyContent: "space-between",
-          alignItems: "center",
-        }}
-      >
-        <div className="listTitle">Task</div>
-        <div>
-          <button
-            id="clearLogBtn"
-            className="clearLogBtn btnStyle ml-15"
-            onClick={handelClearLog}
-            disabled={
-              !getIsStaff() ||
-              logEntriesToDelete.log_ids.length === 0 ||
-              logEntriesToDelete.task_ids.length === 0
-            }
+    <div>
+      {showDhcpTable ? (
+        <div className="listContainer">
+          <div
+            style={{
+              display: "flex",
+              justifyContent: "space-between",
+              alignItems: "center",
+            }}
           >
-            Clear
-          </button>
-
-          <button className="clearLogBtn btnStyle ml-15" onClick={clearFilters}>
-            Clear All Filters
-          </button>
-
-          <button
-            id="refreshLogBtn"
-            className="clearLogBtn btnStyle ml-15"
-            onClick={getLogs}
-            disabled={!getIsStaff()}
-          >
-            Refresh
-          </button>
-        </div>
-      </div>
-      <div style={gridStyle} className="ag-theme-alpine ">
-        <AgGridReact
-          ref={gridRef}
-          rowData={logEntries}
-          columnDefs={colDefs}
-          onRowClicked={(params) => {
-            openLogDetails(params);
-          }}
-          stopEditingWhenCellsLoseFocus={true}
-          onSelectionChanged={onSelectionChanged}
-          rowSelection="multiple"
-          pagination={true}
-          paginationPageSize={50}
-          paginationPageSizeSelector={[50, 100, 150, 200]}
-        />
-      </div>
-
-      {showLogDetails === "setupDialog" && (
-        <SetupLogModal
-          logData={logDetails}
-          onClose={() => setShowLogDetails(false)}
-          onSubmit={() => setShowLogDetails(false)}
-          title="Log Details"
-          id="setupLogDetails"
-        />
-      )}
-      {showLogDetails === "genericDialog" && (
-        <GenericLogModal
-          logData={logDetails}
-          onClose={() => setShowLogDetails(false)}
-          onSubmit={() => setShowLogDetails(false)}
-          title="Log Details"
-          id="genericLogDetails"
-        />
-      )}
-      {showLogDetails === "discoveryDialog" && (
-        <DiscoveryLogModal
-          logData={logDetails}
-          onClose={() => setShowLogDetails(false)}
-          onSubmit={() => setShowLogDetails(false)}
-          title="Log Details"
-          id="discoveryLogDetails"
-        />
-      )}
-
-      {showLogDetails === "deleteDialog" && (
-        <Modal
-          show={true}
-          onClose={() => setShowLogDetails("null")}
-          title=""
-          onSubmit={() => setShowLogDetails("null")}
-        >
-          <div>
-            Task which are in Start state can not be cleared. Revoke them and
-            try again
-            <div
-              style={{
-                marginTop: "10px",
-                display: "flex",
-                justifyContent: "center",
-                gap: "10px",
+            <span className="listTitle">
+              Available SONiC Devices in network
+            </span>
+            <CredentialForm
+              type="status"
+              sendCredentialsToParent={(e) => {
+                setSshData(e);
               }}
-            >
+            />
+            <div>
               <button
-                className="btnStyle"
-                onClick={() => setShowLogDetails("null")}
+                className="btnStyle "
+                onClick={discoverDhcp}
+                disabled={!getIsStaff() || selectedRows.length === 0}
               >
-                OK
+                Discover Device
+              </button>
+              <button
+                className="btnStyle ml-15"
+                onClick={() => {
+                  scanDhcp();
+                }}
+                disabled={!getIsStaff()}
+              >
+                Scan DHCP
               </button>
             </div>
           </div>
-        </Modal>
-      )}
+
+          <div
+            className="datatable resizable mt-15"
+            ref={dhcpTableRef}
+            onMouseMove={handleResizeDhcpTable}
+          >
+            <div style={gridStyleDhcpTable} className="ag-theme-alpine">
+              <AgGridReact
+                ref={gridRefDhcpTable}
+                rowData={dhcpTable}
+                columnDefs={dhcpColumn}
+                defaultColDef={defaultColDef}
+                stopEditingWhenCellsLoseFocus={true}
+                onSelectionChanged={onSelectionChangedDhcp}
+                rowSelection="multiple"
+                suppressRowClickSelection={true}
+              ></AgGridReact>
+            </div>
+          </div>
+        </div>
+      ) : null}
+
+      <div
+        className="logPanel resizable"
+        id="logPanel"
+        ref={logPannelDivRef}
+        onMouseMove={handleResize}
+      >
+        <div
+          className="mb-15"
+          style={{
+            display: "flex",
+            justifyContent: "space-between",
+            alignItems: "center",
+          }}
+        >
+          <div className="listTitle">Task</div>
+          <div>
+            <button
+              id="clearLogBtn"
+              className="clearLogBtn btnStyle ml-15"
+              onClick={handelClearLog}
+              disabled={
+                !getIsStaff() ||
+                logEntriesToDelete.log_ids.length === 0 ||
+                logEntriesToDelete.task_ids.length === 0
+              }
+            >
+              Clear
+            </button>
+
+            <button
+              className="clearLogBtn btnStyle ml-15"
+              onClick={clearFilters}
+            >
+              Clear All Filters
+            </button>
+
+            <button
+              id="refreshLogBtn"
+              className="clearLogBtn btnStyle ml-15"
+              onClick={getLogs}
+              disabled={!getIsStaff()}
+            >
+              Refresh
+            </button>
+          </div>
+        </div>
+        <div style={gridStyle} className="ag-theme-alpine ">
+          <AgGridReact
+            ref={gridRef}
+            rowData={logEntries}
+            columnDefs={colDefs}
+            onRowClicked={(params) => {
+              openLogDetails(params);
+            }}
+            stopEditingWhenCellsLoseFocus={true}
+            onSelectionChanged={onSelectionChanged}
+            rowSelection="multiple"
+            pagination={true}
+            paginationPageSize={50}
+            paginationPageSizeSelector={[50, 100, 150, 200]}
+          />
+        </div>
+
+        {showLogDetails === "setupDialog" && (
+          <SetupLogModal
+            logData={logDetails}
+            onClose={() => setShowLogDetails(false)}
+            onSubmit={() => setShowLogDetails(false)}
+            title="Log Details"
+            id="setupLogDetails"
+          />
+        )}
+        {showLogDetails === "genericDialog" && (
+          <GenericLogModal
+            logData={logDetails}
+            onClose={() => setShowLogDetails(false)}
+            onSubmit={() => setShowLogDetails(false)}
+            title="Log Details"
+            id="genericLogDetails"
+          />
+        )}
+        {showLogDetails === "discoveryDialog" && (
+          <DiscoveryLogModal
+            logData={logDetails}
+            onClose={() => setShowLogDetails(false)}
+            onSubmit={() => setShowLogDetails(false)}
+            title="Log Details"
+            id="discoveryLogDetails"
+          />
+        )}
+
+        {showLogDetails === "deleteDialog" && (
+          <Modal
+            show={true}
+            onClose={() => setShowLogDetails("null")}
+            title=""
+            onSubmit={() => setShowLogDetails("null")}
+          >
+            <div>
+              Task which are in Start state can not be cleared. Revoke them and
+              try again
+              <div
+                style={{
+                  marginTop: "10px",
+                  display: "flex",
+                  justifyContent: "center",
+                  gap: "10px",
+                }}
+              >
+                <button
+                  className="btnStyle"
+                  onClick={() => setShowLogDetails("null")}
+                >
+                  OK
+                </button>
+              </div>
+            </div>
+          </Modal>
+        )}
+      </div>
     </div>
   );
 };
