+++ src/components/logpane/logpane.scss
@@ -1,23 +1,27 @@
 .logPanel {
-    width: "100%";
-    height: "470px";
+  width: "100%";
+  height: "470px";
+  padding: 15px;
+  -webkit-box-shadow: 2px 4px 10px 1px rgba(0, 0, 0, 0.47);
+  box-shadow: 2px 4px 10px 1px rgba(201, 201, 201, 0.47);
+  margin: 10px;
 
-    .stickyButton {
-        position: absolute;
-        z-index: 3;
-        width: calc(100vw - 301px);
-        background-color: white;
-        padding-top: 15px;
-        padding-bottom: 15px;
-        margin-top: -15px;
-    }
-  
-    .icon {
-        display: flex;
-        align-items: center;
-        justify-content: center;
-    }
-    .table {
-        padding-top: 45px;
-    }
+  .stickyButton {
+    position: absolute;
+    z-index: 3;
+    width: calc(100vw - 301px);
+    background-color: white;
+    padding-top: 15px;
+    padding-bottom: 15px;
+    margin-top: -15px;
+  }
+
+  .icon {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+  }
+  .table {
+    padding-top: 45px;
+  }
 }
