+++ src/utils/backend_rest_urls.js
@@ -151,10 +151,6 @@ export function dhcpBackupURL(device_ip) {
   return host_addr + "/files/dhcp/backups?device_ip=" + device_ip;
 }
 
-export function dhcpDeviceListURL() {
-  return host_addr + "/files/dhcp/list";
-}
-
 export function dhcpScanURL() {
   return host_addr + "/files/dhcp/scan";
 }
